backend/requirements.txt
django
psycopg2-binary
python-dotenv
djangorestframework
markdown
django-filter